requirements.txt
# Operator 1 -- Python dependencies
# Core
requests>=2.28
pandas>=1.5
numpy>=1.23
pyarrow>=10.0
pyyaml>=6.0
python-dotenv>=1.0  # .env file support for local development

# Temporal modeling -- Phase 6
# T6.1: regime detection + causality
statsmodels>=0.14
scikit-learn>=1.2
ruptures>=1.1
hmmlearn>=0.3

# T6.2: forecasting models (actively used in forecasting.py)
arch>=6.0
xgboost>=1.7
torch>=2.0

# T6.1: Bayesian change-point detection (used in regime_detector.py)
pymc>=5.0

# Not yet implemented -- uncomment when adding these modules:
# SALib>=1.4          # Sobol sensitivity analysis
# TA-Lib>=0.4         # Candlestick pattern detection
# deap>=1.4           # Genetic algorithm optimisation
# pywt>=1.4           # Wavelet decomposition

# Report generation (Phase 7, used in report_generator.py)
matplotlib>=3.6

# Not yet used:
# plotly>=5.13

# Testing
pytest>=7.0
